requirements.txt
torch
torchaudio
openai-whisper
groq
streamlit
sounddevice
scipy
librosa
pandas